+++ nest-forms-backend/src/admin/admin.module.ts
@@ -4,7 +4,7 @@ import UserInfoPipeModule from '../auth/decorators/user-info-pipe.module'
 import ClientsModule from '../clients/clients.module'
 import ConvertModule from '../convert/convert.module'
 import NasesModule from '../nases/nases.module'
-import NasesUtilsService from '../nases/utils-services/tokens.nases.service'
+import NasesSenderService from '../nases/services/nases.sender.service'
 import PrismaModule from '../prisma/prisma.module'
 import TaxModule from '../tax/tax.module'
 import ThrowerErrorGuard from '../utils/guards/thrower-error.guard'
@@ -20,7 +20,7 @@ import AdminController from './admin.controller'
     UserInfoPipeModule,
     NasesModule,
   ],
-  providers: [ThrowerErrorGuard, NasesUtilsService, MinioClientSubservice],
+  providers: [ThrowerErrorGuard, NasesSenderService, MinioClientSubservice],
   exports: [],
   controllers: [AdminController],
 })
