+++ nest-forms-backend/src/ginis/ginis.service.spec.ts
@@ -13,7 +13,8 @@ import prismaMock from '../../test/singleton'
 import ClientsService from '../clients/clients.service'
 import BaConfigService from '../config/ba-config.service'
 import ConvertService from '../convert/convert.service'
-import NasesUtilsService from '../nases/utils-services/tokens.nases.service'
+import NasesContactsService from '../nases/services/nases.contacts.service'
+import NasesSenderService from '../nases/services/nases.sender.service'
 import PrismaService from '../prisma/prisma.service'
 import MailgunService from '../utils/global-services/mailer/mailgun.service'
 import ThrowerErrorGuard from '../utils/guards/thrower-error.guard'
@@ -91,8 +92,8 @@ describe('GinisService', () => {
         ThrowerErrorGuard,
         { provide: PrismaService, useValue: prismaMock },
         {
-          provide: NasesUtilsService,
-          useValue: createMock<NasesUtilsService>(),
+          provide: NasesSenderService,
+          useValue: createMock<NasesSenderService>(),
         },
         { provide: getQueueToken('sharepoint'), useValue: { add: jest.fn() } },
         {
@@ -128,28 +129,22 @@ describe('GinisService', () => {
       value: { error: jest.fn(), debug: jest.fn(), log: jest.fn() },
     })
 
-    // Create a real NasesUtilsService instance for extraction methods
+    // Create a real NasesContactsService instance for extraction methods
     // The extraction methods are pure functions that don't need dependencies
-    const realNasesUtilsService = new NasesUtilsService(
-      createMock(),
+    const realNasesContactsService = new NasesContactsService(
       module.get(ThrowerErrorGuard),
-      createMock<PrismaService>(),
-      createMock(),
-      createMock(),
-      createMock(),
-      createMock(),
     )
 
     // Use real implementations for extraction methods
     jest
-      .spyOn(service['nasesUtilsService'], 'extractNaturalPersonData')
+      .spyOn(service['nasesContactsService'], 'extractNaturalPersonData')
       .mockImplementation((contact) =>
-        realNasesUtilsService.extractNaturalPersonData(contact),
+        realNasesContactsService.extractNaturalPersonData(contact),
       )
     jest
-      .spyOn(service['nasesUtilsService'], 'extractCorporateBodyData')
+      .spyOn(service['nasesContactsService'], 'extractCorporateBodyData')
       .mockImplementation((contact) =>
-        realNasesUtilsService.extractCorporateBodyData(contact),
+        realNasesContactsService.extractCorporateBodyData(contact),
       )
   })
 
@@ -866,7 +861,7 @@ describe('GinisService', () => {
         .mockResolvedValue('contactId')
 
       jest
-        .spyOn(service['nasesUtilsService'], 'createTechnicalAccountJwtToken')
+        .spyOn(service['nasesSenderService'], 'createTechnicalAccountJwtToken')
         .mockReturnValue('jwt-token')
     })
