+++ nest-forms-backend/src/nases-consumer/nases-consumer.module.ts
@@ -8,7 +8,7 @@ import FormValidatorRegistryModule from '../form-validator-registry/form-validat
 import FormsModule from '../forms/forms.module'
 import GinisModule from '../ginis/ginis.module'
 import NasesModule from '../nases/nases.module'
-import NasesUtilsService from '../nases/utils-services/tokens.nases.service'
+import NasesSenderService from '../nases/services/nases.sender.service'
 import RabbitmqClientModule from '../rabbitmq-client/rabbitmq-client.module'
 import TaxModule from '../tax/tax.module'
 import MailgunService from '../utils/global-services/mailer/mailgun.service'
@@ -36,7 +36,7 @@ import WebhookSubservice from './subservices/webhook.subservice'
   ],
   providers: [
     NasesConsumerService,
-    NasesUtilsService,
+    NasesSenderService,
     ThrowerErrorGuard,
     MailgunHelper,
     MailgunService,
