+++ nest-forms-backend/src/nases/nases.service.spec.ts
@@ -33,7 +33,7 @@ import ThrowerErrorGuard from '../utils/guards/thrower-error.guard'
 import { JwtNasesPayloadDto, UpdateFormRequestDto } from './dtos/requests.dto'
 import { NasesErrorsEnum, NasesErrorsResponseEnum } from './nases.errors.enum'
 import NasesService from './nases.service'
-import NasesUtilsService from './utils-services/tokens.nases.service'
+import NasesSenderService from './services/nases.sender.service'
 
 jest.mock('forms-shared/definitions/getFormDefinitionBySlug')
 jest.mock('forms-shared/form-utils/validators')
@@ -70,8 +70,8 @@ describe('NasesService', () => {
         },
         ThrowerErrorGuard,
         {
-          provide: NasesUtilsService,
-          useValue: createMock<NasesUtilsService>(),
+          provide: NasesSenderService,
+          useValue: createMock<NasesSenderService>(),
         },
         {
           provide: PrismaService,
@@ -150,7 +150,7 @@ describe('NasesService', () => {
   })
 
   describe('sendFormEid', () => {
-    it('should throw an error if sendMessageNases throws', async () => {
+    it('should throw an error if send throws', async () => {
       // Mock dependencies
       const mockForm = {
         id: '1',
@@ -183,7 +183,7 @@ describe('NasesService', () => {
       service['formsService'].checkFormBeforeSending = jest
         .fn()
         .mockResolvedValue(mockForm)
-      service['nasesUtilsService'].createUserJwtToken = jest
+      service['NasesSenderService'].createUserJwtToken = jest
         .fn()
         .mockReturnValue('mock-jwt')
       service['filesService'].areFormAttachmentsReady = jest
@@ -198,7 +198,7 @@ describe('NasesService', () => {
 
       // this is the important mock we're testing against
       jest
-        .spyOn(service['nasesUtilsService'], 'sendMessageNases')
+        .spyOn(service['NasesSenderService'], 'send')
         .mockResolvedValue({ status: 404, data: {} })
 
       const updateFormSpy = jest.spyOn(service['formsService'], 'updateForm')
@@ -313,7 +313,7 @@ describe('NasesService', () => {
         mockFormDefinition,
       )
       jest
-        .spyOn(service['nasesUtilsService'], 'sendMessageNases')
+        .spyOn(service['NasesSenderService'], 'send')
         .mockResolvedValue({ status: 404, data: {} })
       const sendToNasesSpy = jest.spyOn(service, 'sendToNasesAndUpdateState')
 
@@ -356,7 +356,7 @@ describe('NasesService', () => {
         mockFormDefinition,
       )
       jest
-        .spyOn(service['nasesUtilsService'], 'sendMessageNases')
+        .spyOn(service['NasesSenderService'], 'send')
         .mockResolvedValue({ status: 200, data: {} })
       const sendToNasesSpy = jest.spyOn(service, 'sendToNasesAndUpdateState')
       const publishToGinisSpy = jest
@@ -401,7 +401,7 @@ describe('NasesService', () => {
         mockFormDefinition,
       )
       jest
-        .spyOn(service['nasesUtilsService'], 'sendMessageNases')
+        .spyOn(service['NasesSenderService'], 'send')
         .mockResolvedValue({ status: 200, data: {} })
       const sendToNasesSpy = jest.spyOn(service, 'sendToNasesAndUpdateState')
       const publishToGinisSpy = jest.spyOn(
@@ -458,7 +458,7 @@ describe('NasesService', () => {
         mockFormDefinition,
       )
       jest
-        .spyOn(service['nasesUtilsService'], 'sendMessageNases')
+        .spyOn(service['NasesSenderService'], 'send')
         .mockResolvedValue({ status: 200, data: {} })
       const sendToNasesSpy = jest.spyOn(service, 'sendToNasesAndUpdateState')
       const publishToGinisSpy = jest.spyOn(
@@ -673,7 +673,7 @@ describe('NasesService', () => {
 
   describe('sendToNasesAndUpdateState', () => {
     it('should throw if status is not 200', async () => {
-      service['nasesUtilsService'].sendMessageNases = jest
+      service['NasesSenderService'].send = jest
         .fn()
         .mockResolvedValue({ status: 401 })
 
@@ -695,7 +695,7 @@ describe('NasesService', () => {
     })
 
     it('should start checking for nases delivery and not trigger any errors', async () => {
-      service['nasesUtilsService'].sendMessageNases = jest
+      service['NasesSenderService'].send = jest
         .fn()
         .mockResolvedValue({ status: 200 })
 
@@ -718,7 +718,7 @@ describe('NasesService', () => {
     })
 
     it('should pass additionalFormUpdates to formsService.updateForm', async () => {
-      service['nasesUtilsService'].sendMessageNases = jest
+      service['NasesSenderService'].send = jest
         .fn()
         .mockResolvedValue({ status: 200 })
 
@@ -750,7 +750,7 @@ describe('NasesService', () => {
     })
 
     it('should update to ERROR an throw error if sending to NASES fails', async () => {
-      service['nasesUtilsService'].sendMessageNases = jest
+      service['NasesSenderService'].send = jest
         .fn()
         .mockResolvedValue({ status: 500 })
 
@@ -779,7 +779,7 @@ describe('NasesService', () => {
     })
 
     it('should update to DELIVERED_NASES if sending to NASES is successful', async () => {
-      service['nasesUtilsService'].sendMessageNases = jest
+      service['NasesSenderService'].send = jest
         .fn()
         .mockResolvedValue({ status: 200 })
