+++ nest-forms-backend/src/nases/services/nases.sender.service.ts
@@ -1,5 +1,4 @@
-import type { WithImplicitCoercion } from 'node:buffer'
-import * as crypto from 'node:crypto'
+
 import { Stream } from 'node:stream'
 
 import { Injectable } from '@nestjs/common'
@@ -16,10 +15,6 @@ import { extractFormSubjectTechnical } from 'forms-shared/form-utils/formDataExt
 import { buildSlovenskoSkXml } from 'forms-shared/slovensko-sk/xmlBuilder'
 import jwt from 'jsonwebtoken'
 import mime from 'mime-types'
-import {
-  UpvsCorporateBody,
-  UpvsNaturalPerson,
-} from 'openapi-clients/slovensko-sk'
 import { v1 as uuidv1, v4 as uuidv4 } from 'uuid'
 
 import ClientsService from '../../clients/clients.service'
@@ -39,58 +34,11 @@ import { NasesAttachmentXmlObject } from '../dtos/xml.dto'
 import {
   NasesErrorCodesEnum,
   NasesErrorCodesResponseEnum,
-  NasesErrorsEnum,
   NasesErrorsResponseEnum,
 } from '../nases.errors.enum'
 
-export interface NaturalPersonData {
-  given_names?: string[]
-  family_names?: { value?: string; primary?: boolean }[]
-  birth?: {
-    date?: string
-  }
-}
-
-export interface CorporateBodyData {
-  name?: string
-  /** IČO */
-  cin?: string
-  /** DIČ */
-  tin?: string | null
-}
-
-export function isUpvsNaturalPerson(
-  contact: UpvsNaturalPerson | UpvsCorporateBody,
-): contact is UpvsNaturalPerson {
-  return (
-    contact.type === 'natural_person' &&
-    'natural_person' in contact &&
-    contact.natural_person !== undefined
-  )
-}
-
-export function isUpvsCorporateBody(
-  contact: UpvsNaturalPerson | UpvsCorporateBody,
-): contact is UpvsCorporateBody {
-  return (
-    contact.type === 'legal_entity' &&
-    'corporate_body' in contact &&
-    contact.corporate_body !== undefined
-  )
-}
-
-export interface NaturalPersonExtractedData {
-  firstNames: string[]
-  lastNames: string[]
-}
-
-export interface CorporateBodyExtractedData {
-  ico?: string
-  name?: string
-}
-
 @Injectable()
-export default class NasesUtilsService {
+export default class NasesSenderService {
   private readonly logger: LineLoggerSubservice
 
   constructor(
@@ -102,29 +50,7 @@ export default class NasesUtilsService {
     private configService: ConfigService,
     private readonly clientsService: ClientsService,
   ) {
-    this.logger = new LineLoggerSubservice('NasesUtilsService')
-  }
-
-  private tokenValidation(
-    rawHead: string,
-    rawBody: string,
-    signature:
-      | WithImplicitCoercion<string>
-      | {
-          [Symbol.toPrimitive](hint: 'string'): string
-        },
-  ): boolean {
-    const publicKey = this.configService.getOrThrow<string>('OBO_TOKEN_PUBLIC')
-    const verifyObject = crypto.createVerify('RSA-SHA256')
-    verifyObject.write(`${rawHead}.${rawBody}`)
-    verifyObject.end()
-    const base64Signature = Buffer.from(signature, 'base64').toString('base64')
-    const signatureIsValid = verifyObject.verify(
-      publicKey,
-      base64Signature,
-      'base64',
-    )
-    return signatureIsValid
+    this.logger = new LineLoggerSubservice('NasesSenderService')
   }
 
   private async stream2buffer(stream: Stream): Promise<Buffer> {
@@ -475,7 +401,7 @@ export default class NasesUtilsService {
   }
 
   // TODO nicer error handling, for now it is assumed this function never throws and a lot of code relies on that
-  async sendMessageNases(
+  async send(
     jwtToken: string,
     data: Forms,
     senderUri: string,
@@ -551,110 +477,5 @@ export default class NasesUtilsService {
         },
       }
     }
-  }
-
-  async isNasesMessageDelivered(formId: string): Promise<boolean> {
-    const jwtToken = this.createTechnicalAccountJwtToken()
-    const result = await this.clientsService.slovenskoSkApi
-      .apiEdeskMessagesSearchGet(formId, undefined, undefined, {
-        headers: {
-          Authorization: `Bearer ${jwtToken}`,
-        },
-      })
-      .then((response) => response.data.length > 0)
-      .catch((error: unknown) => {
-        this.logger.error(
-          this.throwerErrorGuard.InternalServerErrorException(
-            NasesErrorsEnum.SEND_TO_NASES_ERROR,
-            'Error when checking if message is in eDesk',
-            undefined,
-            error,
-          ),
-        )
-        return false
-      })
-
-    return result
-  }
-
-  extractNaturalPersonData(
-    contact: UpvsNaturalPerson,
-  ): NaturalPersonExtractedData {
-    const naturalPerson = contact.natural_person as
-      | NaturalPersonData
-      | null
-      | undefined
-
-    const result: NaturalPersonExtractedData = {
-      firstNames: [],
-      lastNames: [],
-    }
-
-    if (!naturalPerson) {
-      return result
-    }
-
-    // given_names is already an array of strings, maintain order
-    if (naturalPerson.given_names && naturalPerson.given_names.length > 0) {
-      result.firstNames = [...naturalPerson.given_names]
-    }
-
-    // family_names is an array of objects, extract values maintaining order
-    // Sort by primary first (primary names come first), then maintain array order
-    if (naturalPerson.family_names && naturalPerson.family_names.length > 0) {
-      const sortedFamilyNames = [...naturalPerson.family_names].sort((a, b) => {
-        // Primary names come first
-        if (a.primary && !b.primary) return -1
-        if (!a.primary && b.primary) return 1
-        return 0 // Maintain original order for same primary status
-      })
-      result.lastNames = sortedFamilyNames
-        .map((name) => name.value)
-        .filter((value): value is string => value !== undefined)
-    }
-
-    return result
-  }
-
-  extractCorporateBodyData(
-    contact: UpvsCorporateBody,
-  ): CorporateBodyExtractedData {
-    const corporateBody = contact.corporate_body as
-      | CorporateBodyData
-      | null
-      | undefined
-
-    const result: CorporateBodyExtractedData = {}
-
-    if (corporateBody) {
-      if (corporateBody.name) {
-        result.name = corporateBody.name
-      }
-      if (corporateBody.cin) {
-        result.ico = corporateBody.cin
-      }
-    }
-
-    // Fallback: check various_ids for ICO if not found in corporate_body
-    if (!result.ico && contact.various_ids) {
-      const icoEntry = contact.various_ids.find(
-        (id) =>
-          id.type?.name?.includes('IČO') ||
-          id.type?.name?.includes('Identifikačné číslo organizácie'),
-      )
-      if (icoEntry?.value) {
-        result.ico = icoEntry.value
-      }
-
-      // don't throw, alert only
-      this.logger.error(
-        this.throwerErrorGuard.UnprocessableEntityException(
-          NasesErrorsEnum.IDENTITY_SEARCH_DATA_INCONSISTENT,
-          `extractCorporateBodyData: ${NasesErrorsResponseEnum.IDENTITY_SEARCH_DATA_INCONSISTENT}: ICO not found in contact returned by nases ${contact.uri}.`,
-        ),
-      )
-    }
-
-    return result
   }
 }
